Добавлен отчет за месяц

+++ src/main/java/ru/yandex/finance_tracker/service/AccountService.java
@@ -11,5 +11,5 @@ public interface AccountService {
 
     AccountInfoDto createAccount(Long userId, AccountCreateRequest request);
 
-    List<TransactionInfoDto> getTransactionsByAccountId(Integer accountId,  int page, int size);
+    List<TransactionInfoDto> getTransactionsByAccountId(Long accountId,  int page, int size);
 }
